adds neovim theme
this commit extends the nugu theme with neovim colorscheme support.
to be exact, it does only cooperate with the `nugu` colorscheme in
neovim to create a dark and light variant.

only the palette is exported as a lua file to nvim.
the rest of the plumbing and the definitions of highlight groups are
still in neovim, and will probably stay there.